+++ src/components/meal-card/UpdateFoodFormGroupItem.tsx
@@ -2,7 +2,7 @@ import { useContext } from "react";
 import { ListGroup } from "react-bootstrap";
 import { Action, exitFoodEditModeAction, updateFoodAction } from "../../actions";
 import { Food } from "../../model/Food";
-import { FoodInputForm } from "../FoodInputForm";
+import { FoodInputForm } from "../input-form/FoodInputForm";
 import { MealDispatch } from "../MealDispatch";
 
 export const UpdateFoodFormGroupItem = (props: { food: Food; mealIndex: number; foodIndex: number; }) => {
